api_extract/pyproject.toml
[project]
name = "api-extract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.43",
    "yfinance>=0.2.65",
]
